[VSPC-17] Created Display superclass and handled SpaceDoesNotExistException for spacelinks
