+++ vspace/src/main/java/edu/asu/diging/vspace/core/model/display/ISpaceLinkDisplay.java
@@ -1,21 +1,12 @@
 package edu.asu.diging.vspace.core.model.display;
 
 import edu.asu.diging.vspace.core.model.ISpaceLink;
-import edu.asu.diging.vspace.core.model.IVSpaceElement;
 
-public interface ISpaceLinkDisplay extends IVSpaceElement {
+public interface ISpaceLinkDisplay extends IDisplay {
 
     String getId();
 
     void setId(String id);
-
-    float getPositionX();
-
-    void setPositionX(float positionX);
-
-    float getPositionY();
-
-    void setPositionY(float positionY);
 
     void setLink(ISpaceLink link);
