+++ vspace/src/main/java/edu/asu/diging/vspace/core/model/display/impl/ExternalLinkDisplay.java
@@ -10,19 +10,15 @@ import org.hibernate.annotations.Parameter;
 import edu.asu.diging.vspace.core.model.IExternalLink;
 import edu.asu.diging.vspace.core.model.display.IExternalLinkDisplay;
 import edu.asu.diging.vspace.core.model.impl.ExternalLink;
-import edu.asu.diging.vspace.core.model.impl.VSpaceElement;
 
-public class ExternalLinkDisplay extends VSpaceElement implements IExternalLinkDisplay {
+public class ExternalLinkDisplay extends Display implements IExternalLinkDisplay {
 
     @Id
     @GeneratedValue(generator = "link-display-id-generator")
-    @GenericGenerator(name = "externallink-display-id-generator",
-      parameters = @Parameter(name = "prefix", value = "EXLD"), 
+    @GenericGenerator(name = "externallink-display-id-generator", 
+      parameters = @Parameter(name = "prefix", value = "EXLD"),
       strategy = "edu.asu.diging.vspace.core.data.IdGenerator")
     private String id;
-
-    private float positionX;
-    private float positionY;
 
     @OneToOne(targetEntity = ExternalLink.class)
     private IExternalLink externalLink;
@@ -53,51 +49,7 @@ public class ExternalLinkDisplay extends VSpaceElement implements IExternalLinkD
      * (non-Javadoc)
      * 
      * @see edu.asu.diging.vspace.core.model.display.impl.IExternalLinkDisplay#
-     * getPositionX( )
-     */
-    @Override
-    public float getPositionX() {
-        return positionX;
-    }
-
-    /*
-     * (non-Javadoc)
-     * 
-     * @see edu.asu.diging.vspace.core.model.display.impl.IExternalLinkDisplay#
-     * setPositionX(float)
-     */
-    @Override
-    public void setPositionX(float positionX) {
-        this.positionX = positionX;
-    }
-
-    /*
-     * (non-Javadoc)
-     * 
-     * @see edu.asu.diging.vspace.core.model.display.impl.IExternalLinkDisplay#
-     * getPositionY( )
-     */
-    @Override
-    public float getPositionY() {
-        return positionY;
-    }
-
-    /*
-     * (non-Javadoc)
-     * 
-     * @see edu.asu.diging.vspace.core.model.display.impl.IExternalLinkDisplay#
-     * setPositionY(float)
-     */
-    @Override
-    public void setPositionY(float positionY) {
-        this.positionY = positionY;
-    }
-
-    /*
-     * (non-Javadoc)
-     * 
-     * @see edu.asu.diging.vspace.core.model.display.impl.IExternalLinkDisplay#
-     * getExternalLink(float)
+     * getExternalLink()
      */
     @Override
     public IExternalLink getExternalLink() {
@@ -114,5 +66,4 @@ public class ExternalLinkDisplay extends VSpaceElement implements IExternalLinkD
     public void setExternalLink(IExternalLink externalLink) {
         this.externalLink = externalLink;
     }
-
 }
